refactor: move hover actions and mailToContact
